+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Developers should add to the **Unreleased** section once they feel their branch 
 - Added reps, thanks, "đánh" features
 - Added mock someone feature
 - Added ask 8ball (-8ball) and quote of the day (-qotd) feature
+- Added animated emoji for non Nitro user
 
 ### Changed